requirements.txt
Django>=5.2,<6.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
whitenoise>=6.6,<7.0
dj-database-url>=2.1,<3.0
python-dotenv>=1.0,<2.0
django-globus-portal-framework>=0.4,<1.0
social-auth-app-django>=5.0,<6.0
django-plotly-dash>=2.4,<3.0
django-svelte-jsoneditor>=0.4,<0.5
plotly>=5.0,<6.0
